refactor and add mutexes
