+++ 19_http_server/server.go
@@ -16,7 +16,7 @@ type PlayerServer struct {
 	Store PlayerStore
 }
 
-func (p *PlayerServer) ServeHTTP(w http.ResponseWriter, r *http.Request) { // The Handler interface is what we need to implement in order to make a server, and the HandlerFunc type is an adapter to allow the use of ordinary functions as HTTP handlers. HandlerFunc documentation: https://pkg.go.dev/net/http#HandlerFunc and Handler documentation: https://pkg.go.dev/net/http#Handler
+func (p *PlayerServer) ServeHTTP(w http.ResponseWriter, r *http.Request) { // The http.Handler interface is what we need to implement in order to make a server, and the HandlerFunc type is an adapter to allow the use of ordinary functions as HTTP handlers. http.HandlerFunc documentation: https://pkg.go.dev/net/http#HandlerFunc and http.Handler documentation: https://pkg.go.dev/net/http#Handler
 	player := strings.TrimPrefix(r.URL.Path, "/players/")
 
 	switch r.Method {
