+++ 19_in_memory_player_store.go
@@ -1,0 +1,24 @@
+package main
+
+import "sync"
+
+func NewInMemoryPlayerStore() *InMemoryPlayerStore {
+	return &InMemoryPlayerStore{map[string]int{}, sync.RWMutex{}}
+}
+
+type InMemoryPlayerStore struct {
+	store map[string]int
+	lock  sync.RWMutex
+}
+
+func (i *InMemoryPlayerStore) GetPlayerScore(name string) int {
+	i.lock.RLock()
+	defer i.lock.RUnlock()
+	return i.store[name]
+}
+
+func (i *InMemoryPlayerStore) RecordWin(name string) {
+	i.lock.Lock()
+	defer i.lock.Unlock()
+	i.store[name]++
+}
